refactor: вынести цвета в ресурсы и удалить лишние файлы
