+++ app/src/main/java/ru/goaliepash/playlistmaker/ItunesApi.kt
@@ -3,7 +3,6 @@ package ru.goaliepash.playlistmaker
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Query
-import ru.goaliepash.playlistmaker.model.SearchResponse
 
 interface ItunesApi {
