+++ app/src/main/java/ru/goaliepash/playlistmaker/OnTrackClickListener.kt
@@ -1,7 +1,5 @@
 package ru.goaliepash.playlistmaker
 
-import ru.goaliepash.playlistmaker.model.Track
-
 fun interface OnTrackClickListener {
     fun onTrackClick(track: Track)
 }
