+++ app/src/main/java/ru/goaliepash/playlistmaker/TrackAdapter.kt
@@ -11,7 +11,6 @@ import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.RecyclerView.ViewHolder
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
-import ru.goaliepash.playlistmaker.model.Track
 import java.text.SimpleDateFormat
 import java.util.Locale
