--- app/src/main/res/drawable-night/edit_text_search_background.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-
-    <solid android:color="@color/yp_white" />
-
-    <corners android:radius="8dp" />
-
-</shape>
