--- app/src/main/res/drawable-night/ic_search_no_internet.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="120dp"
-    android:height="120dp"
-    android:viewportWidth="120"
-    android:viewportHeight="120">
-
-    <path
-        android:fillColor="#ffffff"
-        android:fillType="evenOdd"
-        android:pathData="M34,29.5L111,3V31.5L40,42.1V76.5V79H39.89C38.95,89.14 32.42,97 24.5,97C15.94,97 9,87.82 9,76.5C9,65.18 15.94,56 24.5,56C28.08,56 31.38,57.6 34,60.3V43V35V29.5Z" />
-
-    <path
-        android:fillColor="#3772E7"
-        android:pathData="M71.5,62.5m-34.5,0a34.5,34.5 0,1 1,69 0a34.5,34.5 0,1 1,-69 0" />
-
-    <path
-        android:fillColor="#ffffff"
-        android:pathData="M72.16,66.53L86.31,80.69C86.67,81.04 87.15,81.23 87.65,81.23C88.15,81.22 88.63,81.02 88.98,80.67C89.33,80.32 89.53,79.84 89.54,79.34C89.54,78.84 89.35,78.36 89,78L54.73,43.7C54.56,43.51 54.35,43.37 54.11,43.27C53.88,43.17 53.63,43.12 53.38,43.11C53.13,43.11 52.87,43.16 52.64,43.26C52.41,43.35 52.19,43.49 52.01,43.67C51.84,43.85 51.69,44.06 51.6,44.3C51.5,44.53 51.46,44.78 51.46,45.04C51.46,45.29 51.51,45.54 51.61,45.77C51.71,46 51.86,46.21 52.04,46.39L57.94,52.29C56.71,53.11 55.55,54.06 54.47,55.14C53.49,56.13 52.61,57.2 51.82,58.35C51.68,58.55 51.58,58.79 51.53,59.03C51.47,59.28 51.47,59.53 51.51,59.77C51.6,60.27 51.89,60.71 52.3,61C52.51,61.14 52.74,61.25 52.98,61.3C53.23,61.35 53.48,61.36 53.73,61.31C54.22,61.22 54.66,60.94 54.95,60.52C55.64,59.52 56.4,58.6 57.16,57.83C58.22,56.77 59.41,55.84 60.69,55.05L62.89,57.25C60.27,58.75 58.15,60.99 56.79,63.69C56.57,64.13 56.54,64.65 56.7,65.12C56.86,65.59 57.2,65.98 57.65,66.21C58.09,66.43 58.6,66.48 59.08,66.33C59.55,66.18 59.95,65.85 60.18,65.41C60.74,64.3 61.47,63.3 62.34,62.43C63.32,61.45 64.46,60.65 65.71,60.08L68.39,62.76C66.55,63.17 64.87,64.1 63.55,65.43C62.67,66.32 61.96,67.36 61.47,68.51C61.27,68.97 61.27,69.49 61.45,69.95C61.64,70.42 62,70.79 62.46,71C62.92,71.2 63.44,71.21 63.91,71.03C64.37,70.85 64.75,70.49 64.96,70.03C65.28,69.31 65.71,68.66 66.24,68.13C67,67.37 67.94,66.82 68.98,66.54C70.02,66.26 71.12,66.25 72.16,66.53ZM69.62,55.23L73.86,59.47C75.75,60.02 77.46,61.04 78.84,62.43C79.72,63.31 80.45,64.33 81.01,65.44C81.12,65.67 81.27,65.87 81.46,66.04C81.66,66.21 81.88,66.34 82.12,66.42C82.36,66.5 82.61,66.53 82.86,66.51C83.12,66.49 83.36,66.42 83.59,66.31C83.81,66.19 84.01,66.03 84.18,65.84C84.34,65.64 84.46,65.42 84.54,65.18C84.61,64.93 84.64,64.68 84.62,64.43C84.59,64.17 84.52,63.93 84.4,63.71C83.65,62.24 82.69,60.9 81.54,59.73C79.98,58.18 78.12,56.97 76.06,56.19C74.01,55.41 71.81,55.09 69.62,55.23ZM63.88,49.48L67.02,52.62C70.62,51.94 74.35,52.33 77.73,53.73C81.12,55.14 84.03,57.5 86.09,60.53C86.23,60.74 86.41,60.91 86.63,61.05C86.84,61.19 87.07,61.28 87.32,61.33C87.56,61.38 87.82,61.37 88.06,61.32C88.31,61.27 88.54,61.17 88.75,61.03C88.96,60.88 89.13,60.7 89.27,60.49C89.4,60.28 89.5,60.04 89.54,59.79C89.59,59.55 89.58,59.29 89.53,59.05C89.47,58.8 89.37,58.57 89.23,58.36C86.5,54.37 82.57,51.36 78.01,49.76C73.45,48.16 68.5,48.06 63.88,49.48ZM72.54,70.66C72.81,70.92 73.02,71.24 73.16,71.59C73.31,71.93 73.38,72.31 73.38,72.68C73.38,73.06 73.31,73.43 73.16,73.78C73.02,74.12 72.81,74.44 72.54,74.7C72.28,74.97 71.96,75.18 71.62,75.32C71.27,75.47 70.9,75.54 70.52,75.54C70.15,75.54 69.78,75.47 69.43,75.32C69.08,75.18 68.77,74.97 68.5,74.7C67.97,74.17 67.66,73.44 67.66,72.68C67.66,71.92 67.97,71.19 68.5,70.66C69.04,70.12 69.77,69.82 70.52,69.82C71.28,69.82 72.01,70.12 72.54,70.66Z" />
-
-    <path
-        android:fillColor="#ffffff"
-        android:fillType="evenOdd"
-        android:pathData="M111,6L105,18.21V79.3C102.38,76.61 99.08,75 95.5,75C86.94,75 80,84.18 80,95.5C80,106.82 86.94,116 95.5,116C103.68,116 110.38,107.62 110.96,97H111V95.5V6Z" />
-
-</vector>
