+++ app/src/main/res/drawable/edit_text_search_background.xml
@@ -2,7 +2,7 @@
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
     android:shape="rectangle">
 
-    <solid android:color="@color/yp_light_gray" />
+    <solid android:color="@color/search_background" />
 
     <corners android:radius="8dp" />
