+++ app/src/main/res/values-night/colors.xml
@@ -20,4 +20,13 @@
     <color name="play_fill_color">@color/yp_white</color>
     <color name="track_artist_text_color">@color/yp_white</color>
     <color name="track_name_info_text_color">@color/yp_white</color>
+    <color name="search_background">@color/yp_white</color>
+    <color name="search_error_fill_color_1">@color/yp_white</color>
+    <color name="search_error_fill_color_2">@color/yp_black</color>
+    <color name="search_error_fill_color_3">@color/yp_yellow</color>
+    <color name="search_error_fill_color_4">@color/yp_white</color>
+    <color name="search_no_internet_fill_color_1">@color/yp_white</color>
+    <color name="search_no_internet_fill_color_2">@color/yp_blue</color>
+    <color name="search_no_internet_fill_color_3">@color/yp_white</color>
+    <color name="search_no_internet_fill_color_4">@color/yp_white</color>
 </resources>
